refactor nested web-server stack
When running this from scratch there were problems because the service couldn’t be added
to a target group that wasn’t associated with a load balancer.  So the service shouldn’t be created
until the listeners are created. This requires a DependsOn reference, which means the service
and listeners need to be in the same stack.