decrease solr draining time
Solr is configured to only run one instance at a time. This is because the storage for solr is on the container disk.
So if two instances were running then then each would end up with only part of the data.
So the faster we can drain the solr connections and bring up a new instance the better.