Here is an attempt at putting cloudfront infront of LARA
This was to support the sensor-interactive
This approach has problems, https goes into http on LARA
But then LARA doesn't pick up on this proxied protocol.
This is a pretty common problem.  A fix could be in the LARA code.
Or the we might be able to use https between cloudfront and LARA
that might fix the problem.