template to setup generic project resources
For projects that don't need to share URLs to their resources, this approach
saves us a cloudfront distribution and DNS entry.
I'd suspect most projects don't share their resource URLs. These are
just used for images and videos inside of their content.

A project that wants to use this still needs an IAM role and group created
to restrict them to just their project folder. That would be in a new template.
That template could be extracted from the project-resources-with-domain template.