looking at the metrics more I decided to lower the thresholds further
The response time was not getting much better even when the average CPU usage was fairly low.
Hopefully the scale up threshold is low enough now that we will increase the number tasks pretty quickly.
And we'll keep them high for most of the day.

It might be better to increase the step boundary since the Cpu reservation is decreased, we are more likely
to blow through the step boundary pretty much every time there is a scale up alarm.