Extract Portal Params `VpcId` and `SubnetIds`
There were some hardcoded values in the `portal-ecs.yml` and `elastic-search.yml` templates
that prevented them from being used to create stacks in the QA environment.

The templates were updated with new Params that include default values for the production environment
and comments with the correct values for using the templates in the QA enviornment.